Refactor util functions
